requirements.txt
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
idna==2.10
pkg-resources==0.0.0
requests==2.25.1
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.2
soupsieve==2.2.1
urllib3==1.26.5
